IDE betrayed me. Partial revert of 3656340b1d481ed70d504ca6973f0f76de896a98.

+++ src/main/java/li/cil/sedna/riscv/R5CPUTemplate.java
@@ -1944,7 +1944,7 @@ final class R5CPUTemplate implements R5CPU {
                       @Field("rs1") final int rs1,
                       @Field("rs2") final int rs2) {
         if (rd != 0) {
-            x[rd] = (long) (int) x[rs1] << (int) x[rs2];
+            x[rd] = (int) x[rs1] << (int) x[rs2];
         }
     }
 
@@ -2160,7 +2160,7 @@ final class R5CPUTemplate implements R5CPU {
                       @Field("rs1") final int rs1,
                       @Field("rs2") final int rs2) {
         if (rd != 0) {
-            x[rd] = (long) (int) x[rs1] * (int) x[rs2];
+            x[rd] = (int) x[rs1] * (int) x[rs2];
         }
     }
